Use CURRENTFUNC() macro, be more consistent (and verbose) with types

+++ tests/behaviour/contracts/cross_contract_calls/invalid_output.sol
@@ -13,10 +13,10 @@ interface I{
 contract A{
     /// warp-cairo
     /// DECORATOR(external)
-    /// func invalid_bde9e40e() -> (number : felt) {
+    /// func CURRENTFUNC()() -> (number : felt) {
     ///     return (783495,);
     /// }
-    function invalid() public returns (uint) {
+    function invalid() public returns (uint128) {
     }
 }
